Add NetHack smoke test and SDK I/O fixes
Make test-nethack reach the NetHack login prompt by returning a static SDK passwd entry for icsos/UID 0, fixing legacy FILE fread item counts, kernel fgets partial-line EOF behavior, and ELF64 stream page-read comparison. Verified test-nethack, test-integration, test-iobench, test-posixio, test-spawn, and test-dup. test-fork remains a pre-existing failure at HEAD e075877 (GPF64 TSS-bad in fork_child_return).

+++ ics-os/kernel/module/elf_module.c
@@ -932,9 +932,10 @@ int elf64_stream_load(char *module_name, int mode, char *p, char *workdir,
                mapok = 0;
                break;
             }
-            if (fread(pagebuf, (int)len, 1, f) != (int)len) {
+            /* One fseek+fread per page: len <= 4096 covers the whole page. */
+            if (fread(pagebuf, (size_t)len, 1, f) != (int)len) {
                printf("elf64-stream: fread fail %s off=%llu len=%llu\n",
-                      module_name, fileoff, len);
+                      module_name, fileoff, (unsigned long long)len);
                mapok = 0;
                break;
             }
